Add in a new Element::Get method.
This one is a little bit easier to use in that it returns
just a boolean, and the value is stuffed into _param, a
reference.  This comes out of the discussion here:
https://bitbucket.org/osrf/sdformat/pull-requests/333 .

While I was in here, I ended up sharing more of the logic
for ::Get so we have less duplicated code.  I also added
tests for all of this.