Switch to using cstdint in Exception.
According to https://stackoverflow.com/questions/13642827/cstdint-vs-stdint-h ,
using cstdint should be preferred over using stdint.h.  The former puts
things in the std:: namespace and optionally in the global namespace,
while the latter puts things only in the global namespace.  Thus,
in order to be maximally compliant with the spec and with namespaces,
use cstdint and always refer to int64_t explicitly in the std::
namespace.