Enforce minimum/maximum values specified in SDFormat description files (Retry PR) (#303)
Currently, some SDFormat description files contain min and max values,
but the parser does not enforce those values in any way.
This PR adds a validation step in the parser that checks if scalar values
are within the allowed range.
Note that this PR addresses only element values, not attributes.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>