Add Model::CanonicalLinkAndRelativeName
This is a new private function that de-duplicates the
code that was in both FrameSemantics.cc and Model::CanonicalLink.
It provides a Link* pointer to the canonical link and its
nested name relative to the current model, which is needed
in the FrameAttachedToGraph.
This function is private to hide :: syntax from libsdformat9,
at least until there is a compelling reason to make the API
public.
A helper function is added to FrameSemantics.cc as a friend
of Model so that buildFrameAttachedToGraph can call the
private API. That function can't be added directly
as a friend since it uses a FrameAttachedToGraph&
as an argument, which isn't defined in Model.hh.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>