Add cmake option to disable console logfile
The sdf::Console class logs some messages to a file
in ~/.sdformat/sdformat.log by default.
This wraps the code in the constructor of the sdf::Console
class that opens this file in an `#ifndef` and adds
a cmake option `SDFORMAT_DISABLE_CONSOLE_LOGFILE`
to easily disable the use of this logfile.

Initial approach to #334.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>