Return positive infinity instead of -1 for boundless positive values (#357)
SDFormat has a convention of using -1 to represent positive infinity
for symmetric joint limits on velocity and effort.
If users of the DOM API are not careful and cognizant of this
convention, it can result in confusing bugs that are hard to track down.

This PR changes the API behavior so that the API returns positive
infinity instead of -1 to represent boundlessly positive values.
Hopefully this might save other SDF users from long, confusing
debugging sessions.

A test is added with an example file with default, finite and infinite
joint limit values, using both "inf" and "-1" to specify infinite values
of the effort and velocity limit.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>

Co-authored-by: Michael X. Grey <grey@openrobotics.org>