Handle placement frames with new PoseRelativeTo construction scheme
The PoseRelativeTo used to be constructed in each nested model's Load
funcition. Now, it's only constructed at the outer most model. Because
of this decoupling and Since the model is `const` when the graph is
constructed, placement frames are handled differently. Instead of
updating the raw pose of the model, the edge connecting the model frame
to it's relative_to frame is modified to take into account the placement
frame.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>