Add __root__ scope
The __root__ vertex is the root node of both world and model
PoseRelativeTo graphs. It corresponds to the `<sdf>` tag in SDFormat
files. Having this node makes it possible to keep the a model's pose and
(possibly accounting for placement_frame) information in the edge from
the _root__ vertex to the model vertex.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>