Revert the SemanticPose::Resolve API that takes SemanticPose objects
This proved to be problematic since the graph contained in the input
SemanticPose can be at a different scope than the graph in the current
object. Thus it becomes necessary to find the least common ancestor
vertex between two vertices in the graph. This is feasible, but it would
be best to do it in a separate PR.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>