Make sdf::Model::CanonicalLinkAndRelativeName public
This member function is the only way to obtain the relative name of the
canonical link when the link is implicitly nested, i.e, the //model/@canonical_link
attribute is empty and the canonical link resolves to a link nested in a
child model. While `sdf::Model::CanonicalLink` can be used to resolve
the link, the returned `sdf::Link*` can only provide the local
(non-relative) name of the link.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>