Add ParserConfig class to encapsulate file path settings (#439)
This adds a new ParserConfig object that contains configuration options for the libsdformat parser. This replaces the old globals g_uriPathMap and g_findFileCB. To do this in a backward compatible way, the ParserConfig class provides a singleton object that calls to sdf::setSdfFindCallback and sdf::addURIPath use.

New overloads of the main SDFormat file and string loading APIs allow users to pass in a ParserConfig object. If omitted, these APIs will default to the singleton ParserConfig object.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>

Co-authored-by: Steve Peters <scpeters@openrobotics.org>