Translate poses of nested models inside other nested models
When addNestedModel function is called, it processes links and joints
to translate the pose accordingly to parent pose. This was not done
for nested models inside the SDF being processed.

The change includes nested models in the same way that is doing for
links and add a test that fails without the change.

Signed-off-by: Jose Luis Rivero <jrivero@osrfoundation.org>