Replace custom cmake code with ign-cmake2 (#780)
This replaces most of the custom cmake code in
libsdformat with the functionality provided by
ignition-cmake2. The root CMakeLists.txt is much shorter
now and most of the cmake folder has been deleted.
This is made possible by the NO_IGNITION_PREFIX
and REPLACE_IGNITION_INCLUDE_PATH parameters
added to ign_configure_project in ign-cmake#190 and
ign-cmake#191. Closes #181. Other details:

* Use FindIgnURDFDOM from ign-cmake#193
* Use HIDE_SYMBOLS_BY_DEFAULT from ign-cmake#196
* Set LEGACY_PROJECT_PREFIX from ign-cmake#199

Signed-off-by: Steve Peters <scpeters@openrobotics.org>