ParserConfig: add ToElementUseIncludeTag policy
This adds a new policy to the ParserConfig class to replace
the boolean parameter in the Model, Root, and World
ToElement methods that specifies whether include tags should
be used or the complete model. ParserConfig objects are then
passed to the ToElement methods instead of the bool,
allowing the ParserConfig to be used in additional
sdf::initFile API calls. This changes only unreleased
parts of the API, which is permitted.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>