Revert some instances of `sdf::ParserConfig` propgation
The only use of `sdf::ParserConfig` in `sdf::initFile` and its callees
is when searching for the description (.sdf) file on disk using
`sdf::findFile`. However, all the files used within the `ToElement`
calls are standard `.sdf` files that are kepts in an embedded map. i.e.,
`sdf::findFile` is not needed to find them. Therefore, we can use the
default `sdf::initFile` in these functions.

There is no harm in passing along the `sdf::ParserConfig` object, but it
would be inconsistent with the rest of `*::ToElement` functions which do
not receive a config object.

The same applies to the change in `sdf::Root::Load`.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>