inertial.sdf spec: improve documentation
This helps resolve ambiguities in the SDF <inertial> tag by
clarifying the link’s center of mass location, explicitly
states that the inertia is always about the link’s center of
mass, clarifies directions associated with moments/products of
inertia, and clarifies the ± sign convention for product of inertia.

Thanks to Paul Mitiguy and Michael Sherman from Toyota Research
Institute for these improvements.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>