Updated findfile() to search localpath first (#1292)
As the issue above mentions, this tutorial for Garden, is not able to parse the SDF file if it was named model.sdf. It works if the file name is changed. This was happening due to the search order used in the sdf::findFile() function. The used search order was causing the findFile() to find the model.sdf from the install path instead of the one present locally. Since the found model.sdf is actually the SDF spec file for the <model> and it is not a valid SDFormat file, it caused the script to throw a not valid error.
This PR changes the search order used in the findFile() function by searching for the local path (current directory) first and solves the issue.
---------

Signed-off-by: Jasmeet Singh <jasmeet0915@gmail.com>
Co-authored-by: Addisu Z. Taddese <addisu@openrobotics.org>