Fix version variable in reference manual PDF filename
The variable SDF_VERSION_FULL is a C preprocessor macro which is set to
the value of the CMake variable PROJECT_VERSION_FULL in config.hh.in. It
is not defined as a CMake variable itself, and currently evaluates to an
empty string resulting in refman.pdf being renamed to sdf-.pdf.

Signed-off-by: Scott K Logan <logans@cottsay.net>