Change behavior of Param::Get<bool>
Previously when a Param was set from a string, the
`Get<bool>` method would always return `true`, and
the value would be `true` if the lowercase value string
matched `"1"` or `"true"`. Otherwise the value would
be `false`.

Now, the `Get<bool>` method returns `true` only if the
lowercase value string matches one of `"0"`, `"1"`,
`"true"`, or `"false"`. Otherwise `Get<bool>` returns
`false` and the value is not written.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>