Allow empty strings in plugin and custom attributes (#1407)
Currently errors are generated when adding an attribute
containing an empty string to a <plugin> block or as a
custom attribute. This adds failing tests to confirm the bug
and fixes the errors in by setting `_required == 0` when
calling Element::AddAttribute. This also changes
Element::ToString to print empty custom attributes.

Fixes #725.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>