Disable latex and class hierarchy generation
We don't use the latex output anywhere, but it brings in a big
dependency when building documentation or deb packages.
The class hierarchy is also not that useful since there is not much
inheritance used in the codebase. This also brings the doxygen
parameters of sdformat closer to other Gazebo libraries.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>