Fix some cmake logic related to python (#1557)
* Remove unneeded check for python3 from sdf folder

We already require Python3, so remove the check
and build error from sdf/CMakeLists.txt

* Fix check for psutil python package

Search for python psutil package even if
BUILD_TESTING is not set and move error message
to the place where the memory leak test is
disabled if the package is missing. This is
needed because the BUILD_TESTING variable is
not defined until after gz_configure_build.

Signed-off-by: Steve Peters <scpeters@openrobotics.org>