Backport load optimization without deprecation
This is a partial backport of PR 1589 without the
deprecations. Original PR description:

Optimize load time and memory consumption by reusing element descriptions (#1589)

In most usages of `GetElementDescription`, the returned `sdf::ElementPtr` is not mutated, thus it is possible to implement a copy-on-write mechanism to avoid unnecessary clones of element descriptions. This PR removes clones of element description during the loading of an SDFormat file and introduces APIs to get immutable element descriptions. It also provides mutable element descriptions which clone the element description when called.

* Do not clone element descriptions
* Add MutableElementDescription and deprecate GetElementDescription
* Add tests
* Narrower use of GZ_UTILS_WARN_* macros
* Reword comments, Migration guide
* Remove duplicate line in python test
* Add / adjust //// lines

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>
Signed-off-by: Steve Peters <scpeters@openrobotics.org>
Co-authored-by: Steve Peters <scpeters@openrobotics.org>