Manual forward port of: Bazel updates (#1597) (#1599)
The change had to be amended to apply it on main. Specifically, the repo archive_overrides in MODULE.bazel for gz deps was removed in that PR on the Jetty branch, but we want to preserve it on main to ensure CI uses gz deps from HEAD.

-- Original PR description
Few small fixes in MODULE.bazel as pre-work to add automation to push new releases to BCR.

Drop repo_name, which removes the need to patch MODULE.bazel when pushing a release to BCR. repo_name is not a required field and can be added on the client side during import if needed to disambiguate packages.
Add compatibility_level to match what is set in BCR
Bump buildifier_prebuilt since version 8.2.0 has a bug in bazel 8 presubmits. Also remove unnecessary buildifier targets in test/BUILD.bazel; the top-level targets in BUILD.bazel cover the whole repo.

Signed-off-by: Shameek Ganguly <shameek@intrinsic.ai>