Introduce a cache to improve performance of construction of SDFormat
This commit should reduce load times significantly as it caches the first parse of the spec
so that it does not need to be reparsed at library load time through static initiallization of a pass.

On my m4 macbook air this reduces the following benchmark:

void deserialize()
{
    std::string sdf = "<?xml version=\"1.0\" ?><sdf version='1.6'></sdf>";
    sdf::Root root;
    sdf::Errors errors = root.LoadSdfString(sdf);
    if (!root.Model())
    {
        return;
    }
    sdf::Model _model = *root.Model();
    return;
}

int main(int argc, char **argv)
{
    std::chrono::steady_clock::time_point begin = std::chrono::steady_clock::now();
    for (int i = 0; i < 100; i++)
    deserialize();
    std::chrono::steady_clock::time_point end = std::chrono::steady_clock::now();

    std::cout << "Whole function call took: " <<
        std::chrono::duration_cast<std::chrono::microseconds>(end - begin).count()
        << "[µs]" << std::endl;
}

from running for 300000µs to 3000µs.

Signed-off-by: Arjo Chakravarty <arjo@openrobotics.org>